requirements.txt
openai
python-dotenv
textstat